+++ CHANGELOG.md
@@ -514,3 +514,4 @@
 - 2026-04-26: Features mod, send, summaries
 - 2026-04-27: Docs summaries, mod, send
 - 2026-04-28: Features mod, summaries
+- 2026-04-29: Async mod, summaries